refactor(auth): 패키지 구조 개선으로 전역 모듈과 도메인 모듈 구분 명확화 (#6, #11) - 요청 객체 오타 수정 포함
