+++ src/main/java/com/example/wait4eat/domain/auth/controller/AuthController.java
@@ -1,10 +1,10 @@
-package com.example.wait4eat.global.auth.controller;
+package com.example.wait4eat.domain.auth.controller;
 
-import com.example.wait4eat.global.auth.dto.request.SigninRequest;
-import com.example.wait4eat.global.auth.dto.request.SignupRequest;
-import com.example.wait4eat.global.auth.dto.response.SigninResponse;
-import com.example.wait4eat.global.auth.dto.response.SignupResponse;
-import com.example.wait4eat.global.auth.service.AuthService;
+import com.example.wait4eat.domain.auth.dto.request.SigninRequest;
+import com.example.wait4eat.domain.auth.dto.request.SignupRequest;
+import com.example.wait4eat.domain.auth.dto.response.SigninResponse;
+import com.example.wait4eat.domain.auth.dto.response.SignupResponse;
+import com.example.wait4eat.domain.auth.service.AuthService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
